[Concurrency] Encode deferred task storage as a flag
Move the current-task storage kinds into a public C-compatible debug ABI header and reserve the high bit of the storage-kind byte for platform-deferred resolution.\n\nHave each Threading implementation declare its storage kind explicitly. Embedded sets the deferred flag, while linked platform libraries publish a concrete storage kind through the debug ABI symbol.
