+++ include/swift/Threading/Impl/ThreadEmbeddedImpl.h
@@ -31,7 +31,8 @@
 #define SWIFT_THREADING_USE_RESERVED_TLS_KEYS 1
 #define SWIFT_THREADING_HAS_LAZY_MUTEX 0
 #define SWIFT_THREADING_HAS_CONDITION_VARIABLE 0
-#define SWIFT_THREADING_PLATFORM_DEFINED 1
+#define SWIFT_THREAD_LOCAL_STORAGE_KIND                                        \
+  SWIFT_CONCURRENCY_CURRENT_TASK_STORAGE_KIND_DEFERRED_MASK
 
 static_assert(SWIFT_TLS_KEY_COUNT ==
                   static_cast<int>(swift::tls_key::exclusivity) + 1,
